copilot fixes
