+++ benchmarks/benchmarks/bench_fftnd.py
@@ -4,21 +4,9 @@ import numpy as np
 
 import mkl_fft
 
+from ._utils import _make_input
+
 _RNG_SEED = 42
-
-
-def _make_input(rng, shape, dtype):
-    """Return an array of the given *shape* and *dtype*.
-
-    Complex dtypes are populated with non-zero imaginary parts so the
-    benchmark exercises a genuine complex transform path.
-    """
-    dt = np.dtype(dtype)
-    if dt.kind == "c":
-        return (
-            rng.standard_normal(shape) + 1j * rng.standard_normal(shape)
-        ).astype(dt)
-    return rng.standard_normal(shape).astype(dt)
 
 
 # ---------------------------------------------------------------------------
